Split checkout mutations into separate files
