+++ saleor/graphql/checkout/mutations/checkout_add_promo_code.py
@@ -1,0 +1,108 @@
+import graphene
+from django.core.exceptions import ValidationError
+
+from ....checkout.error_codes import CheckoutErrorCode
+from ....checkout.fetch import (
+    fetch_checkout_info,
+    fetch_checkout_lines,
+    update_delivery_method_lists_for_checkout_info,
+)
+from ....checkout.utils import add_promo_code_to_checkout
+from ...core.descriptions import DEPRECATED_IN_3X_INPUT
+from ...core.mutations import BaseMutation
+from ...core.scalars import UUID
+from ...core.types.common import CheckoutError
+from ...core.validators import validate_one_of_args_is_in_mutation
+from ..types import Checkout
+from .utils import (
+    get_checkout_by_token,
+    update_checkout_shipping_method_if_invalid,
+    validate_checkout_email,
+)
+
+
+class CheckoutAddPromoCode(BaseMutation):
+    checkout = graphene.Field(
+        Checkout, description="The checkout with the added gift card or voucher."
+    )
+
+    class Arguments:
+        checkout_id = graphene.ID(
+            description=(
+                f"The ID of the checkout. {DEPRECATED_IN_3X_INPUT} Use token instead."
+            ),
+            required=False,
+        )
+        token = UUID(description="Checkout token.", required=False)
+        promo_code = graphene.String(
+            description="Gift card code or voucher code.", required=True
+        )
+
+    class Meta:
+        description = "Adds a gift card or a voucher to a checkout."
+        error_type_class = CheckoutError
+        error_type_field = "checkout_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info, promo_code, checkout_id=None, token=None):
+        # DEPRECATED
+        validate_one_of_args_is_in_mutation(
+            CheckoutErrorCode, "checkout_id", checkout_id, "token", token
+        )
+
+        if token:
+            checkout = get_checkout_by_token(token)
+        # DEPRECATED
+        else:
+            checkout = cls.get_node_or_error(
+                info, checkout_id or token, only_type=Checkout, field="checkout_id"
+            )
+
+        validate_checkout_email(checkout)
+
+        manager = info.context.plugins
+        discounts = info.context.discounts
+
+        lines, unavailable_variant_pks = fetch_checkout_lines(checkout)
+        if unavailable_variant_pks:
+            not_available_variants_ids = {
+                graphene.Node.to_global_id("ProductVariant", pk)
+                for pk in unavailable_variant_pks
+            }
+            raise ValidationError(
+                {
+                    "lines": ValidationError(
+                        "Some of the checkout lines variants are unavailable.",
+                        code=CheckoutErrorCode.UNAVAILABLE_VARIANT_IN_CHANNEL.value,
+                        params={"variants": not_available_variants_ids},
+                    )
+                }
+            )
+
+        shipping_channel_listings = checkout.channel.shipping_method_listings.all()
+        checkout_info = fetch_checkout_info(
+            checkout, lines, discounts, manager, shipping_channel_listings
+        )
+
+        add_promo_code_to_checkout(
+            manager,
+            checkout_info,
+            lines,
+            promo_code,
+            discounts,
+        )
+
+        update_delivery_method_lists_for_checkout_info(
+            checkout_info,
+            checkout_info.checkout.shipping_method,
+            checkout_info.checkout.collection_point,
+            checkout_info.shipping_address,
+            lines,
+            discounts,
+            manager,
+            shipping_channel_listings,
+        )
+
+        update_checkout_shipping_method_if_invalid(checkout_info, lines)
+        manager.checkout_updated(checkout)
+        return CheckoutAddPromoCode(checkout=checkout)
