+++ saleor/graphql/checkout/mutations/checkout_shipping_method_update.py
@@ -1,0 +1,209 @@
+from typing import Optional
+
+import graphene
+from django.core.exceptions import ValidationError
+
+from ....checkout.error_codes import CheckoutErrorCode
+from ....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
+from ....checkout.utils import (
+    delete_external_shipping_id,
+    is_shipping_required,
+    recalculate_checkout_discount,
+    set_external_shipping_id,
+)
+from ....plugins.webhook.utils import APP_ID_PREFIX
+from ....shipping import interface as shipping_interface
+from ....shipping import models as shipping_models
+from ....shipping.utils import convert_to_shipping_method_data
+from ...core.descriptions import DEPRECATED_IN_3X_INPUT
+from ...core.mutations import BaseMutation
+from ...core.scalars import UUID
+from ...core.types.common import CheckoutError
+from ...core.utils import from_global_id_or_error
+from ...core.validators import validate_one_of_args_is_in_mutation
+from ...shipping.types import ShippingMethod
+from ..types import Checkout
+from .utils import ERROR_DOES_NOT_SHIP, clean_delivery_method, get_checkout_by_token
+
+
+class CheckoutShippingMethodUpdate(BaseMutation):
+    checkout = graphene.Field(Checkout, description="An updated checkout.")
+
+    class Arguments:
+        checkout_id = graphene.ID(
+            description=(
+                f"The ID of the checkout. {DEPRECATED_IN_3X_INPUT} Use token instead."
+            ),
+            required=False,
+        )
+        token = UUID(description="Checkout token.", required=False)
+        shipping_method_id = graphene.ID(required=True, description="Shipping method.")
+
+    class Meta:
+        description = "Updates the shipping method of the checkout."
+        error_type_class = CheckoutError
+        error_type_field = "checkout_errors"
+
+    @staticmethod
+    def _resolve_delivery_method_type(id_) -> Optional[str]:
+        if id_ is None:
+            return None
+
+        possible_types = ("ShippingMethod", APP_ID_PREFIX)
+        type_, id_ = from_global_id_or_error(id_)
+        str_type = str(type_)
+
+        if str_type not in possible_types:
+            raise ValidationError(
+                {
+                    "shipping_method_id": ValidationError(
+                        "ID does not belong to known shipping methods",
+                        code=CheckoutErrorCode.INVALID.value,
+                    )
+                }
+            )
+
+        return str_type
+
+    @classmethod
+    def perform_mutation(
+        cls, _root, info, shipping_method_id, checkout_id=None, token=None
+    ):
+        # DEPRECATED
+        validate_one_of_args_is_in_mutation(
+            CheckoutErrorCode, "checkout_id", checkout_id, "token", token
+        )
+
+        if token:
+            checkout = get_checkout_by_token(token)
+        # DEPRECATED
+        else:
+            checkout = cls.get_node_or_error(
+                info, checkout_id or token, only_type=Checkout, field="checkout_id"
+            )
+
+        manager = info.context.plugins
+
+        lines, unavailable_variant_pks = fetch_checkout_lines(checkout)
+        if unavailable_variant_pks:
+            not_available_variants_ids = {
+                graphene.Node.to_global_id("ProductVariant", pk)
+                for pk in unavailable_variant_pks
+            }
+            raise ValidationError(
+                {
+                    "lines": ValidationError(
+                        "Some of the checkout lines variants are unavailable.",
+                        code=CheckoutErrorCode.UNAVAILABLE_VARIANT_IN_CHANNEL.value,
+                        params={"variants": not_available_variants_ids},
+                    )
+                }
+            )
+
+        checkout_info = fetch_checkout_info(
+            checkout, lines, info.context.discounts, manager
+        )
+        if not is_shipping_required(lines):
+            raise ValidationError(
+                {
+                    "shipping_method": ValidationError(
+                        ERROR_DOES_NOT_SHIP,
+                        code=CheckoutErrorCode.SHIPPING_NOT_REQUIRED.value,
+                    )
+                }
+            )
+
+        type_name = cls._resolve_delivery_method_type(shipping_method_id)
+
+        if type_name == "ShippingMethod":
+            return cls.perform_on_shipping_method(
+                info, shipping_method_id, checkout_info, lines, checkout, manager
+            )
+        return cls.perform_on_external_shipping_method(
+            info, shipping_method_id, checkout_info, lines, checkout, manager
+        )
+
+    @staticmethod
+    def _check_delivery_method(
+        checkout_info,
+        lines,
+        *,
+        delivery_method: Optional[shipping_interface.ShippingMethodData],
+    ) -> None:
+        delivery_method_is_valid = clean_delivery_method(
+            checkout_info=checkout_info,
+            lines=lines,
+            method=delivery_method,
+        )
+        if not delivery_method_is_valid or not delivery_method:
+            raise ValidationError(
+                {
+                    "shipping_method": ValidationError(
+                        "This shipping method is not applicable.",
+                        code=CheckoutErrorCode.SHIPPING_METHOD_NOT_APPLICABLE.value,
+                    )
+                }
+            )
+
+    @classmethod
+    def perform_on_shipping_method(
+        cls, info, shipping_method_id, checkout_info, lines, checkout, manager
+    ):
+        shipping_method = cls.get_node_or_error(
+            info,
+            shipping_method_id,
+            only_type=ShippingMethod,
+            field="shipping_method_id",
+            qs=shipping_models.ShippingMethod.objects.prefetch_related(
+                "postal_code_rules"
+            ),
+        )
+        delivery_method = convert_to_shipping_method_data(
+            shipping_method,
+            shipping_models.ShippingMethodChannelListing.objects.filter(
+                shipping_method=shipping_method,
+                channel=checkout_info.channel,
+            ).first(),
+        )
+
+        cls._check_delivery_method(
+            checkout_info, lines, delivery_method=delivery_method
+        )
+
+        delete_external_shipping_id(checkout=checkout)
+        checkout.shipping_method = shipping_method
+        checkout.save(
+            update_fields=["private_metadata", "shipping_method", "last_change"]
+        )
+
+        recalculate_checkout_discount(
+            manager, checkout_info, lines, info.context.discounts
+        )
+        manager.checkout_updated(checkout)
+        return CheckoutShippingMethodUpdate(checkout=checkout)
+
+    @classmethod
+    def perform_on_external_shipping_method(
+        cls, info, shipping_method_id, checkout_info, lines, checkout, manager
+    ):
+        delivery_method = manager.get_shipping_method(
+            checkout=checkout,
+            channel_slug=checkout.channel.slug,
+            shipping_method_id=shipping_method_id,
+        )
+
+        cls._check_delivery_method(
+            checkout_info, lines, delivery_method=delivery_method
+        )
+
+        set_external_shipping_id(checkout=checkout, app_shipping_id=delivery_method.id)
+        checkout.shipping_method = None
+        checkout.save(
+            update_fields=["private_metadata", "shipping_method", "last_change"]
+        )
+
+        recalculate_checkout_discount(
+            manager, checkout_info, lines, info.context.discounts
+        )
+        manager.checkout_updated(checkout)
+        return CheckoutShippingMethodUpdate(checkout=checkout)
