+++ saleor/graphql/checkout/tests/deprecated/test_checkout_shipping_method_update.py
@@ -34,7 +34,10 @@ MUTATION_UPDATE_SHIPPING_METHOD = """
 """
 
 
-@patch("saleor.graphql.checkout.mutations.clean_delivery_method")
+@patch(
+    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
+    "clean_delivery_method"
+)
 def test_checkout_shipping_method_update_by_id(
     mock_clean_shipping,
     staff_api_client,
@@ -79,7 +82,10 @@ def test_checkout_shipping_method_update_by_id(
     assert checkout.shipping_method == shipping_method
 
 
-@patch("saleor.graphql.checkout.mutations.clean_delivery_method")
+@patch(
+    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
+    "clean_delivery_method"
+)
 def test_checkout_shipping_method_update_by_token(
     mock_clean_shipping,
     staff_api_client,
@@ -123,7 +129,10 @@ def test_checkout_shipping_method_update_by_token(
     assert checkout.shipping_method == shipping_method
 
 
-@patch("saleor.graphql.checkout.mutations.clean_delivery_method")
+@patch(
+    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
+    "clean_delivery_method"
+)
 def test_checkout_shipping_method_update_neither_token_and_id_given(
     mock_clean_shipping, staff_api_client, checkout_with_item, shipping_method
 ):
@@ -139,7 +148,10 @@ def test_checkout_shipping_method_update_neither_token_and_id_given(
     assert data["errors"][0]["code"] == CheckoutErrorCode.GRAPHQL_ERROR.name
 
 
-@patch("saleor.graphql.checkout.mutations.clean_delivery_method")
+@patch(
+    "saleor.graphql.checkout.mutations.checkout_shipping_method_update."
+    "clean_delivery_method"
+)
 def test_checkout_shipping_method_update_both_token_and_id_given(
     mock_clean_shipping, staff_api_client, checkout_with_item, shipping_method
 ):
